add select() tests
select() function returns an integer value representing a symbol in the
reel. The selection is done considering the number of instances of each
symbols in the reel.

For example

Reel 1 has 3 symbols:
 - 10 bananas
 - 10 kiwis
 - 10 oranges

the random number will be between 1 and 30 (the sum of the intances)
the selection will be between 0 and 2 (the index of symbols).

Assuming r=15 the selection will be 1 (the second symbol)

See the select.spec.ts test "simple test" for a working example.